requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
dependency-injector==4.41.0
exceptiongroup==1.1.3
fastapi==0.104.0
h11==0.14.0
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
minio==7.1.17
peewee==3.17.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
python-dotenv==1.0.0
python-linq==2.0.4
python-multipart==0.0.6
PyYAML==6.0.1
rich==13.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
